SkinCancer-main/deploy_backend/server/requirements.txt
matplotlib
numpy
tensorflow-addons
tensorflow-model-optimization
tensorflow
ipython
PyYAML
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.13.0